requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Twilio SDK
twilio>=8.11.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.7.4
pydantic-settings>=2.6.0

# Form Data
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-vertexai>=2.0.0
langchain-google-genai>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud & AI
google-cloud-aiplatform>=1.38.0
google-genai>=0.3.0

# Optional for production
# redis>=5.0.0  # For Redis checkpointer
